Cease pinning CI Flutter to minor version number
Instead, the CI will now upgrade Flutter to latest stable version (as long as it’s 3.x). This is a better trade-off for a sample repo. We don’t need to update several lines in 3 separate YAML files every 3 months. Since this is a small sample repo, it is very unlikely that a minor stable release will break anything.